data_streamer/requirements.txt
kafka-python==2.0.2
numpy==1.24.3
pandas==2.0.3
python-dateutil==2.8.2